move print_site_verification_meta to SV module
