+++ includes/Core/Authentication/Authentication.php
@@ -206,13 +206,6 @@ final class Authentication {
 				return $this->authentication_admin_notices( $notices );
 			}
 		);
-
-		$print_site_verification_meta = function() {
-			$this->print_site_verification_meta();
-		};
-
-		add_action( 'wp_head', $print_site_verification_meta );
-		add_action( 'login_head', $print_site_verification_meta );
 	}
 
 	/**
@@ -515,40 +508,6 @@ final class Authentication {
 		$data['moduleToSetup'] = $module_to_setup;
 
 		return $data;
-	}
-
-	/**
-	 * Prints site verification meta in wp_head().
-	 *
-	 * @since 1.0.0
-	 *
-	 * @global wpdb $wpdb WordPress database abstraction object.
-	 */
-	private function print_site_verification_meta() {
-		global $wpdb;
-
-		// Get verification meta tags for all users.
-		$verification_tags = $this->verification_tag->get_all();
-
-		if ( empty( $verification_tags ) ) {
-			return;
-		}
-
-		$allowed_html = array(
-			'meta' => array(
-				'name'    => array(),
-				'content' => array(),
-			),
-		);
-
-		foreach ( $verification_tags as $verification_tag ) {
-			$verification_tag = html_entity_decode( $verification_tag );
-			if ( 0 !== strpos( $verification_tag, '<meta ' ) ) {
-				$verification_tag = '<meta name="google-site-verification" content="' . esc_attr( $verification_tag ) . '">';
-			}
-
-			echo wp_kses( $verification_tag, $allowed_html );
-		}
 	}
 
 	/**
