+++ tests/phpunit/integration/Modules/Site_VerificationTest.php
@@ -11,6 +11,7 @@
 namespace Google\Site_Kit\Tests\Modules;
 
 use Google\Site_Kit\Context;
+use Google\Site_Kit\Core\Authentication\Authentication;
 use Google\Site_Kit\Core\Modules\Module_With_Scopes;
 use Google\Site_Kit\Modules\Site_Verification;
 use Google\Site_Kit\Tests\Core\Modules\Module_With_Scopes_ContractTests;
@@ -47,6 +48,42 @@ class Site_VerificationTest extends TestCase {
 			apply_filters( 'googlesitekit_auth_scopes', array() )
 		);
 		$this->assertTrue( has_action( 'admin_init' ) );
+	}
+
+	/**
+	 * @dataProvider data_register_head_verification_tags
+	 */
+	public function test_register_head_verification_tags( $saved_tag, $expected_output ) {
+		remove_all_actions( 'wp_head' );
+		remove_all_actions( 'login_head' );
+		$site_verification = new Site_Verification( new Context( GOOGLESITEKIT_PLUGIN_MAIN_FILE ) );
+		$site_verification->register();
+
+		set_transient( 'googlesitekit_verification_meta_tags', array( $saved_tag ) );
+
+		$this->assertContains(
+			$expected_output,
+			$this->capture_action( 'wp_head' )
+		);
+
+		$this->assertContains(
+			$expected_output,
+			$this->capture_action( 'login_head' )
+		);
+	}
+
+	public function data_register_head_verification_tags() {
+		return array(
+			array( // Full meta tag stored.
+				'<meta name="google-site-verification" content="test-verification-content">',
+				'<meta name="google-site-verification" content="test-verification-content">',
+			),
+			array(
+				// Only verification token stored.
+				'test-verification-content-2',
+				'<meta name="google-site-verification" content="test-verification-content-2">',
+			),
+		);
 	}
 
 	public function test_get_module_scopes() {
